autoBQ/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-bigquery
pandas
google-cloud-bigquery-datatransfer